+++ README.md
@@ -50,3 +50,9 @@
     - [x] 상하좌우, 대각선 방향으로 한 칸 움직일 수 있다.
     - [x] 피스를 뛰어 넘어 움직일 수 없다.
     - [x] 이동할 위치에 상대피스가 있으면 잡을 수 있다.
+
+# 게임 진행
+- [ ] 게임 시작 안내 문구 출력
+- [ ] 게임 조작법 안내 문구 출력
+- [ ] 사용자 입력에 따라 게임을 진행한다.
+  - [ ] 게임 종료 커맨드 입력 전까지 반복한다.
