+++ src/main/java/chess/domain/PieceSign.java
@@ -1,14 +1,14 @@
-package chess;
+package chess.domain;
 
-import chess.piece.Bishop;
-import chess.piece.Color;
-import chess.piece.EmptyPiece;
-import chess.piece.King;
-import chess.piece.Knight;
-import chess.piece.Pawn;
-import chess.piece.Piece;
-import chess.piece.Queen;
-import chess.piece.Rook;
+import chess.domain.piece.Bishop;
+import chess.domain.piece.Color;
+import chess.domain.piece.EmptyPiece;
+import chess.domain.piece.King;
+import chess.domain.piece.Knight;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Queen;
+import chess.domain.piece.Rook;
 
 public enum PieceSign {
