+++ src/main/java/chess/domain/piece/King.java
@@ -1,8 +1,8 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.Rule;
-import chess.position.Position;
-import chess.position.PositionDifference;
+import chess.domain.Rule;
+import chess.domain.position.Position;
+import chess.domain.position.PositionDifference;
 
 public class King extends Piece {
