+++ src/main/java/chess/domain/piece/Knight.java
@@ -1,8 +1,8 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.Rule;
-import chess.position.Position;
-import chess.position.PositionDifference;
+import chess.domain.Rule;
+import chess.domain.position.Position;
+import chess.domain.position.PositionDifference;
 
 public class Knight extends Piece {
