+++ src/main/java/chess/domain/position/RankDifference.java
@@ -1,4 +1,4 @@
-package chess.position;
+package chess.domain.position;
 
 public class RankDifference extends Difference {
