+++ src/test/java/chess/SpaceTest.java
@@ -3,14 +3,16 @@ package chess;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import chess.piece.Color;
-import chess.piece.EmptyPiece;
-import chess.piece.Pawn;
-import chess.piece.Piece;
-import chess.piece.Rook;
-import chess.position.File;
-import chess.position.Position;
-import chess.position.Rank;
+import chess.domain.PieceSign;
+import chess.domain.Space;
+import chess.domain.piece.Color;
+import chess.domain.piece.EmptyPiece;
+import chess.domain.piece.Pawn;
+import chess.domain.piece.Piece;
+import chess.domain.piece.Rook;
+import chess.domain.position.File;
+import chess.domain.position.Position;
+import chess.domain.position.Rank;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
