+++ src/test/java/chess/position/FileTest.java
@@ -2,6 +2,8 @@ package chess.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import chess.domain.position.File;
+import chess.domain.position.FileDifference;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
